+++ packages/jest-core/src/runJest.ts
@@ -15,10 +15,10 @@ import {
   formatTestResults,
   makeEmptyAggregatedTestResult,
 } from '@jest/test-result';
+import type {Test} from '@jest/test-result';
 import type TestSequencer from '@jest/test-sequencer';
 import type {Config} from '@jest/types';
 import type {ChangedFiles, ChangedFilesPromise} from 'jest-changed-files';
-import type {Test} from 'jest-runner';
 import type {Context} from 'jest-runtime';
 import {requireOrImportModule, tryRealpath} from 'jest-util';
 import {JestHook, JestHookEmitter} from 'jest-watcher';